requirements.txt
anyio==4.9.0
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
config==0.5.1
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycoingecko==3.2.0
python-dotenv==1.0.1
python-telegram-bot==22.0
requests==2.32.3
sniffio==1.3.1
tzlocal==5.3.1
urllib3==2.3.0
Werkzeug==3.1.3
